Add tests for selection of chain and resnum in PDB

+++ AdaptivePELE/tests/testAdaptiveSampling.py
@@ -131,13 +131,16 @@ class TestadaptiveSampling(unittest.TestCase):
             goldenClusters.append(cluster)
         try:
             self.integrationTest(controlFile, goldenPath, outputPath, goldenClusters)
-        except SystemExit:
-            # Catch error for not having PELE installed
-            print ("Warning! There was a sysExit in test3, this is usually "
-                   "caused by not having PELE installed, so it can be ignored "
-                   "if the test are not running on MareNostrum or life")
-            shutil.rmtree(outputPath)
-            shutil.rmtree("tmp_tests_data_3ptb_data_Test3/")
+        except SystemExit as e:
+            if "No trajectories to cluster!" in e.message:
+                # Catch error for not having PELE installed
+                print ("Warning! There was a sysExit in test3, this is usually "
+                    "caused by not having PELE installed, so it can be ignored "
+                    "if the test are not running on MareNostrum or life")
+                shutil.rmtree(outputPath)
+                shutil.rmtree("tmp_tests_data_3ptb_data_Test3/")
+            else:
+                raise e
 
     def testRestartEmptyClustering(self):
         controlFile = "tests/data/3ptb_data/restartTest.conf"
@@ -151,11 +154,14 @@ class TestadaptiveSampling(unittest.TestCase):
             adaptiveSampling.main(controlFile)
             # Assertions
             self.assertTrue(adaptiveSampling.checkIntegrityClusteringObject(clusteringObjectPath))
-        except SystemExit:
-            # Catch error for not having PELE installed
-            print ("Warning! There was a sysExit in test3, this is usually "
-                   "caused by not having PELE installed, so it can be ignored "
-                   "if the test are not running on MareNostrum or life")
+        except SystemExit as e:
+            if "No trajectories to cluster!" in e.message:
+                # Catch error for not having PELE installed
+                print ("Warning! There was a sysExit in test3, this is usually "
+                    "caused by not having PELE installed, so it can be ignored "
+                    "if the test are not running on MareNostrum or life")
+            else:
+                raise e
         # Remove clustering object from the simulation and create and empty one
         open(clusteringObjectPath, "w").close()
         # cleanup
